服务注册、服务端集成及请求处理&&超时处理

+++ SuRPC/codec/codec.go
@@ -7,7 +7,7 @@ type Header struct {
 	// ServiceMethod request method
 	ServiceMethod string
 	// Seq sequence number of request
-	Seq int
+	Seq uint64
 	Err string
 }
